% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmacology-package.R
\docType{package}
\name{pharmacology-package}
\alias{pharmacology}
\alias{pharmacology-package}
\title{pharmacology: Research Compendium of the GIPSO's B. Rhythms and Pharmacology SQLR}
\description{
An R package with the research compendium of the GIPSO's biological rhythms and pharmacology systematic quantitative literature review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipso.github.io/pharmacology/}
  \item \url{https://github.com/gipso/pharmacology/}
  \item Report bugs at \url{https://github.com/gipso/pharmacology/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Alisson da Silva Rodrigues dos Santos \email{alisson.rodriguesds@gmail.com} (\href{https://orcid.org/0000-0002-5455-3562}{ORCID}) [copyright holder]
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
