% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.chronopharmacology-package.R
\docType{package}
\name{rc.chronopharmacology-package}
\alias{rc.chronopharmacology}
\alias{rc.chronopharmacology-package}
\title{rc.chronopharmacology: Research Compendium of the GIPERBIO Chronopharmacology SQLR}
\description{
Research compendium of the GIPERBIO chronopharmacology systematic quantitative literature review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/giperbio/rc.chronopharmacology/}
  \item Report bugs at \url{https://github.com/giperbio/rc.chronopharmacology/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

Authors:
\itemize{
  \item Alisson da Silva Rodrigues dos Santos \email{alisson.rodriguesds@gmail.com} (\href{https://orcid.org/0000-0002-5455-3562}{ORCID}) [copyright holder]
}

}
\keyword{internal}
