% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.chronopharmacology-package.R
\docType{package}
\name{rc.chronopharmacology-package}
\alias{rc.chronopharmacology}
\alias{rc.chronopharmacology-package}
\title{rc.chronopharmacology: Research Compendium of the GIPSO Chronopharmacology SQLR}
\description{
An R package with the research compendium of the GIPSO chronopharmacology systematic quantitative literature review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipso.github.io/rc.chronopharmacology/}
  \item \url{https://github.com/gipso/rc.chronopharmacology/}
  \item Report bugs at \url{https://github.com/gipso/rc.chronopharmacology/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Alisson da Silva Rodrigues dos Santos \email{alisson.rodriguesds@gmail.com} (\href{https://orcid.org/0000-0002-5455-3562}{ORCID}) [copyright holder]
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
